refactor varifocal
