+++ mmyolo/models/losses/__init__.py
@@ -1,4 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .iou_loss import IoULoss, bbox_overlaps
+from .varifocal_loss import varifocal_loss
 
-__all__ = ['IoULoss', 'bbox_overlaps']
+__all__ = ['IoULoss', 'bbox_overlaps', 'varifocal_loss']
